Enable simple replacement for Microsoft.WebApplication.targets (#89)
* Enable simple replacement for Microsoft.WebApplication.targets

Right now, there is a requirement to use msbuild due to the need of Microsoft.WebApplication.targets. This adds an override so that it can be replaced externally to this project.

`WebApplicationsTargetPath` property contains the full path to a .targets file to import.
Defaults to `$(VSToolsPath)\WebApplications\Microsoft.WebApplication.targets` if not set.