Update files to remove explicit version number in readmes and templates (#63)
Add code to replace version number placeholders with current version during build.

Previously, it was difficult to ensure that the templates and readme files had the correct version number included in the examples.
Often it was required to do a build and publish a set of packages to determine the next version number, then update them and publish a new set of packages with just the version numbers updated.

This change now uses placeholders and code to replace them on the fly whilst building the packages.